fix rlimit handling
Commit f0308436b5b6df00f12aef7f54e6f35948a87f84 changed the semantics of rlimit.
Before the commit, rlimit == 0 means no limit, after the commit, rlimit == 0
means limit.

However, the rest of the code has not been updated which the changed semantics.

There is also a related issue that limit is sometimes `unsigned` and sometimes
`uint64_t`, which is not the same type on my architecture.

This commit is a hack to bring everything into norm. However, I'm not sure this
is correct or the best solution. This does, however, fixes the problem of always
getting `(tactic-exception "max. resource limit exceeded")` in release mode on
Linux when compiled with clang-10.

Here is the compilation command that exhibits the issue:

```
cmake -DCMAKE_BUILD_TYPE=Debug -DCMAKE_C_COMPILER=clang-10 -DCMAKE_CXX_COMPILER=clang++-10 -GNinja -DCMAKE_EXE_LINKER_FLAGS=-fuse-ld=lld -DCMAKE_MODULE_LINKER_FLAGS=-fuse-ld=lld -DCMAKE_SHARED_LINKER_FLAGS=-fuse-ld=lld ../ -GNinja -DCMAKE_BUILD_WITH_INSTALL_RPATH=ON -DCMAKE_INSTALL_PREFIX=$(pwd)/run
```

+++ src/cmd_context/context_params.h
@@ -26,7 +26,7 @@ class context_params {
     void set_bool(bool & opt, char const * param, char const * value);
     void set_uint(unsigned & opt, char const * param, char const * value);
 
-    unsigned    m_rlimit;
+    uint64_t    m_rlimit;
 
 public:
     bool        m_auto_config;
@@ -45,10 +45,10 @@ public:
     unsigned    m_timeout;
     bool        m_statistics;
 
-    unsigned rlimit() const { return m_rlimit; }
+    uint64_t rlimit() const { return m_rlimit; }
     context_params();
     void set(char const * param, char const * value);
-    void set_rlimit(unsigned lim) { m_rlimit = lim; }
+    void set_rlimit(uint64_t lim) { m_rlimit = lim; }
     void updt_params();
     void updt_params(params_ref const & p);
     static void collect_param_descrs(param_descrs & d);
